app: move GET /api/status to app.routers.status
